Fix sway user services
This fixes a situation where systemd services would not be restarted
after a sway session ended due to the two having completely different
scopes. We could therefore be in a situation where SWAYSOCK was wrong
because the user services were using the socket id of the old session.
By moving everything into sway, we do not have that problem.

This commit also fixes a long standing issue with the clipboard manager.
It can now take into account when we are trying to access the password
manager's passwords.